Update reference solution for cell tests
This test includes an interface neumann condition. This condition was applied twice because the two sided
skeleton flag was marked as true as well as the skeleton contribution was contributing to the two sides of
the intersection. This results in a double accumulation of the actual flux. And since this reference
solutions are nothing more than a results of previous version of the program, they were wrong.
Hence, this commit simply updates the to the current code. Qualitatively, the effects of the boundaries
are be reduced which agrees with the change mentioned above.