Fix GCC template specialization
The template specialization of the PkFiniteElementMap was failing in GCC because
the order argument (k) was not the same type of the actual map type. Normally there
is an implicit conversion but for some reason it didn't happen here. There is still
to check whether this was a GCC bug, an clang bug (for doing the conversion),
or a standard undefined behavior. Now however, it should be fixed.