Use bit-flags for stages of the model setup
The idea here is to set a flag for each stage of the model setup and a policy that sets the required stages for a certain configuration. For example, the `write` policy requires the model to configure the stages of grid function space, coefficient vector and writer.
Scape writer configuration if "writer" section is not present in config

In case the ini file does not contain the "writer" section then, the whole writer setup is skipped and no error will be shown.