Add model time stepper
This time stepper is inspired on the boost odeit steppers signatures. To be able to use it, the current models need to be modified so that it exports operators and states can be exchanged easily.
The advantage of this approach is that time evolution can be fine tuned by users at will. Moreover, it allows to have several valid states on the same model and be separately advanced in time.