Set all cmake falges in only one file
The dune base image usually read all the options files in the /duneci/cmake-flags/ folder, but exactly because it reads every file the flags are set depending on the order of the filers. The easiest solution is to set all flags into only one file. Then last flags will have precedence.