Refactor to simplify code and improve performance
With the new developments in dune-pdelab, most of the tricks to have multiple domains are not needed anymore. Even more, because of native support of multiple domains, the ordering is able to avoid non-fixed size offsets and binding is much more faster. Now, multi domain should have almost the same performance as single domain cases.

This commit also refactors the parser to be exchangeable with different backends. This adds the possibility to be SBML compatible on the math expressions.

The subspace machinery in pdelab was also improved so interpolations with (sub) grid functions and VTK writing is much more simpler.

Because code is simpler, we do not need the data context nor the factory classes anymore.

The model became completely stateless and now is just an interpreter of the state.

This commit updates the required dune version to 2.8.

Finite volumes are not supported anymore. This is for two reasons: (i) the finite element map needs to be templetized to work with the different grid types on the multidomain case, thus, the virtual wrappers won't work. And (ii) although a one could find a solution, maintaining mixed grid formulations has proven hard to maintain in dune-copasi as well as in the spatial-model-editor.

There is no instantiation of the model into a library. This simplified the CMake files. Now we use and export the dune-copasi target directly instead of using internal targets.

Compilation times and memory requirements dramatically dropped because of less complex code.