Add empty compartments to the basis tree
Now, all compartments have a node in the basis tree even if no components (or scalar_fields) are in the sub-tree. Or in other words, the dynamic degree of those nodes is 0. In practice, this means that boundary conditions (i.e. outflow) with empty compartments will be taken into account for the residual and jacobian assembly.