Use generic extra flux instead of advective velocity
Not every additional flux (wrt than diffusive flux) needs to advective. Thus, here we change to specify a flux which is much more useful and generic than an advective flux. For example, the richards equation has an extra gravity driven flux which is not advective. Additionally, advective fluxes can be easily represented with the new format.