Move cell data out of local equations
Local equations contain the keys and the data used to refer to its own values in the parsers, hence, there is no need to hold the cell data within it. The local operator can update its values manually. This allows to split the cell data object from the parsers and the local equations while still having the correct updates on the local operator.